+++ pnet-data-api-java/src/main/java/at/porscheinformatik/happyrest/apache/ApacheRestCallFactory.java
@@ -19,7 +19,8 @@ import at.porscheinformatik.happyrest.RestParser;
 import at.porscheinformatik.happyrest.RestUtils;
 import at.porscheinformatik.happyrest.jackson.JacksonBasedFormatter;
 import at.porscheinformatik.happyrest.jackson.JacksonBasedParser;
-import at.porscheinformatik.happyrest.util.BinaryParser;
+import at.porscheinformatik.happyrest.util.ByteArrayParser;
+import at.porscheinformatik.happyrest.util.StringParser;
 import at.porscheinformatik.happyrest.util.TextPlainFormatter;
 
 /**
@@ -33,7 +34,7 @@ public class ApacheRestCallFactory implements RestCallFactory
     public static ApacheRestCallFactory create(RestLoggerAdapter loggerAdapter, ObjectMapper mapper)
     {
         RestFormatter formatter = RestFormatter.of(new JacksonBasedFormatter(mapper), new TextPlainFormatter());
-        RestParser parser = RestParser.of(new JacksonBasedParser(mapper), BinaryParser.INSTANCE);
+        RestParser parser = RestParser.of(new JacksonBasedParser(mapper), StringParser.INSTANCE, ByteArrayParser.INSTANCE);
 
         return new ApacheRestCallFactory(loggerAdapter, formatter, parser);
     }