+++ CHANGELOG.md
@@ -3,6 +3,7 @@ pnet-data-api 1.14.6
 
 * Fix missing content type in Java and Apache client.
 * Removed UTF-8 from JSON content type, as it is the default.
+* Added CharArrayParser and StringParser to the REST adapter.
 
 pnet-data-api 1.14.5
 ====================